# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allizo Service API",
        "description": "API documentation for Allizo Service platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://allizo.example.com",
            "description": "Production Server"
        },
        {
            "url": "http://alizo.test/",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/allizo/carts": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get all carts",
                "description": "Display a listing of the resource.\n /**",
                "operationId": "265ef9aaa14c5916a5af1a5cc3c3d083",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "No carts found"
                    }
                }
            }
        },
        "/api/allizo/carts/{id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get a specific cart",
                "operationId": "b79ee94fa5aa1e76ee0ddf90d6d1d07f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart retrieved"
                    },
                    "404": {
                        "description": "Cart not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "Update a cart",
                "description": "Update the specified resource in storage.\n/**",
                "operationId": "e06cfc26390ff7750f962fbdfa6aa1f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart updated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Delete a cart",
                "description": "Remove the specified resource from storage.\n\n /**",
                "operationId": "70dc29c19cc5264a990587ed31fc265c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart deleted"
                    },
                    "404": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/api/allizo/carts/{cartId}/service": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Add a service to a cart",
                "operationId": "a9118f3de4c5ec96d0c338e586aadbeb",
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service added"
                    },
                    "404": {
                        "description": "Cart or service not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Remove a service from a cart",
                "operationId": "5b8fca0cac26bb321242bebfd53ce253",
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service removed"
                    },
                    "404": {
                        "description": "Cart or service not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}